Fix map deletion and useAllDocs (#86)
If we want side panels in multiple windows to be consistent (open same document with same changes), then we must use a BroadcastChannelNetworkAdapter. But that adapter will restore deleted docs, so we must also broadcast deletes. We can use the same broadcast mechanism to delete map activations to keep the panel UIs in sync.